Implement lazy VMCallerCheckedAnyfunc initialization.
Currently, in the instance initialization path, we build an "anyfunc"
for every imported function and every function in the module. These
anyfuncs are used as function references in tables, both within a module
and across modules (via function imports).

Building all of these is quite wasteful if we never use most of them.
Instead, it would be better to build only the ones that we use.

This commit implements a technique to lazily initialize them: at the
point that a pointer to an anyfunc is taken, we check whether it has
actually been initialized. If it hasn't, we do so. This is a bit tricky
to do in a threadsafe way; we do it by using atomics, denoting an
uninitialized entry by a zero in one field (the function pointer
itself), and carefully writing that field last. Multiple concurrent
initialization is a benign race, since the writes are idempotent.

Keeping the information around that is needed for this lazy init
required a bit of refactoring in the InstanceAllocationRequest; a few
fields now live in an `Arc` held by the instance while it exists. The
signature building/handling could probably also be made more efficient,
by building it only once for a `CompiledModule` and sharing it, but the
crate layering (`CompiledModule` is in `jit` which doesn't know about
the engine's shared signature registry) makes this a bit tricky.