memfd: use mapped memfd image only for nonzero part of heap.
It turns out that it's much faster to fault in an anonymous page than a
CoW'd page of zeroes, because the latter requires a copy while the
former can use a pre-zeroed page (which the kernel constructs as an
idle-time hobby).

Thanks for @koute's comment here [1] for leading to this realization in
this context and @alexcrichton for help tracking it down!

[1]
https://github.com/koute/wasmtime/blob/30fc9256560f4f49e26d57bd3ddecb5490e9a811/crates/runtime/src/mmap.rs#L423-L428